move services into the new module
