+++ frontend/src/modules/MLStudio/components/index.js
@@ -1,0 +1,2 @@
+export * from './MLStudioListItem';
+export * from './MLStudioOverview';
