+++ frontend/src/modules/MLStudio/services/createSagemakerStudioUser.js
