+++ frontend/src/modules/MLStudio/views/MLStudioCreateForm.js
@@ -29,12 +29,9 @@ import {
   useSettings
 } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
-import {
-  createSagemakerStudioUser,
-  listEnvironmentGroups,
-  listEnvironments,
-  useClient
-} from 'services';
+import { listEnvironmentGroups, listEnvironments, useClient } from 'services';
+
+import { createSagemakerStudioUser } from '../services';
 
 const MLStudioCreateForm = (props) => {
   const navigate = useNavigate();
