+++ frontend/src/modules/MLStudio/views/MLStudioList.js
@@ -20,8 +20,9 @@ import {
   useSettings
 } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
-import { listSagemakerStudioUsers, useClient } from 'services';
-import MLStudioListItem from '../components/MLStudioListItem';
+import { useClient } from 'services';
+import { listSagemakerStudioUsers } from '../services';
+import { MLStudioListItem } from '../components/';
 
 function MLStudioPageHeader() {
   return (
