+++ frontend/src/services/graphql/SagemakerStudio/index.js
@@ -1,5 +1,5 @@
-export * from './createSagemakerStudioUser';
+export * from '../../../modules/MLStudio/services/createSagemakerStudioUser';
 export * from './deleteSagemakerStudioUser';
 export * from './getSagemakerStudioUser';
 export * from './getSagemakerStudioUserPresignedUrl';
-export * from './listSagemakerStudioUsers';
+export * from '../../../modules/MLStudio/services/listSagemakerStudioUsers';
